Delegate scanner scan-orchestration helpers to registers module
