+++ custom_components/thessla_green_modbus/registers/schema.py
@@ -20,15 +20,7 @@ from __future__ import annotations
 
 import logging
 import re
-
-try:
-    from enum import StrEnum
-except ImportError:  # pragma: no cover - Python < 3.11
-    from enum import Enum
-
-    class StrEnum(str, Enum):  # type: ignore[no-redef]
-        """Compatibility StrEnum fallback for Python < 3.11."""
-
+from enum import StrEnum
 from typing import Any, Literal, cast
 
 import pydantic
