+++ custom_components/thessla_green_modbus/scanner/__init__.py
@@ -1,0 +1,5 @@
+"""Scanner package public exports."""
+
+from .core import DeviceCapabilities, ThesslaGreenDeviceScanner, is_request_cancelled_error
+
+__all__ = ["DeviceCapabilities", "ThesslaGreenDeviceScanner", "is_request_cancelled_error"]
